+++ app/src/main/kotlin/jp/co/yumemi/android/codecheck/RepositoryDetailFragment.kt
@@ -14,7 +14,7 @@ import androidx.compose.ui.platform.ViewCompositionStrategy
 import androidx.fragment.app.Fragment
 import androidx.navigation.fragment.findNavController
 import androidx.navigation.fragment.navArgs
-import jp.co.yumemi.android.codecheck.designsystem.MainTheme
+import jp.co.yumemi.android.codecheck.core.designsystem.MainTheme
 import jp.co.yumemi.android.codecheck.detail.RepositoryDetailScreen
 
 class RepositoryDetailFragment : Fragment() {
