+++ app/src/main/kotlin/jp/co/yumemi/android/codecheck/RepositorySearchFragment.kt
@@ -13,7 +13,7 @@ import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import androidx.navigation.fragment.findNavController
 import dagger.hilt.android.AndroidEntryPoint
 import jp.co.yumemi.android.codecheck.core.data.model.RepositoryItem
-import jp.co.yumemi.android.codecheck.designsystem.MainTheme
+import jp.co.yumemi.android.codecheck.core.designsystem.MainTheme
 import jp.co.yumemi.android.codecheck.search.RepositorySearchScreen
 import jp.co.yumemi.android.codecheck.search.RepositorySearchViewModel
