+++ feature/search/src/main/java/jp/co/yumemi/android/codecheck/search/RepositorySearchScreen.kt
@@ -47,7 +47,7 @@ import androidx.compose.ui.unit.sp
 import jp.co.yumemi.android.codecheck.core.data.model.DataLoadingState
 import jp.co.yumemi.android.codecheck.core.data.model.ErrorType
 import jp.co.yumemi.android.codecheck.core.data.model.RepositoryItem
-import jp.co.yumemi.android.codecheck.designsystem.MainTheme
+import jp.co.yumemi.android.codecheck.core.designsystem.MainTheme
 import jp.co.yumemi.android.codecheck.ui.OwnerIcon
 import kotlinx.coroutines.coroutineScope
